refactor: #77 move interfaces to types/
